+++ Sources/VCore/Data & Domain Helpers/Image Store & Repository/Cache (Memory)/MockImageMemoryCache.swift
@@ -7,7 +7,11 @@
 
 #if DEBUG
 
-import Foundation
+#if canImport(UIKit)
+import UIKit
+#elseif canImport(AppKit)
+import AppKit
+#endif
 
 /// Mock image memory cache.
 nonisolated public final class MockImageMemoryCache: ImageMemoryCacheProtocol {