+++ Sources/VCore/Data & Domain Helpers/Image Store & Repository/Progress Cache (Memory)/ImageProgressMemoryCacheProtocol.swift
@@ -5,7 +5,11 @@
 //  Created by Vakhtang Kontridze on 16/5/26.
 //
 
-import Foundation
+#if canImport(UIKit)
+import UIKit
+#elseif canImport(AppKit)
+import AppKit
+#endif
 
 /// Image progress memory cache protocol.
 nonisolated public protocol ImageProgressMemoryCacheProtocol: AnyObject, Sendable {